+++ src/test/java/org/folio/processing/mapping/InstanceMappingTest.java
@@ -30,6 +30,7 @@ import org.folio.ContributorType;
 import org.folio.Identifier;
 import org.folio.IdentifierType;
 import org.folio.Instance;
+import org.folio.InstanceDateType;
 import org.folio.InstanceType;
 import org.folio.Subject;
 import org.folio.SubjectSource;
@@ -65,6 +66,8 @@ public class InstanceMappingTest {
   private static final String BIB_WITH_REPEATED_020_SUBFIELDS = "src/test/resources/org/folio/processing/mapping/instance/ISBN.mrc";
   private static final String BIB_WITH_REPEATED_600_SUBFIELDS = "src/test/resources/org/folio/processing/mapping/instance/6xx_subjects.mrc";
   private static final String BIB_WITH_REPEATED_600_SUBFIELD_AND_EMPTY_INDICATOR = "src/test/resources/org/folio/processing/mapping/instance/6xx_subjects_without_indicators.mrc";
+  private static final String BIB_WITH_008_DATE = "src/test/resources/org/folio/processing/mapping/instance/008_date.mrc";
+
 
   private static final String BIB_WITH_RESOURCE_TYPE_SUBFIELD_VALUE = "src/test/resources/org/folio/processing/mapping/instance/336_subfields_mapping.mrc";
   private static final String BIB_WITH_720_FIELDS = "src/test/resources/org/folio/processing/mapping/instance/720_fields_samples.mrc";
@@ -78,6 +81,8 @@ public class InstanceMappingTest {
   private static final String DEFAULT_RESOURCE_IDENTIFIERS_TYPES_PATH = "src/test/resources/org/folio/processing/mapping/instance/resourceIdentifiers.json";
   private static final String DEFAULT_SUBJECT_SOURCES_PATH = "src/test/resources/org/folio/processing/mapping/instance/subjectSources.json";
   private static final String DEFAULT_SUBJECT_TYPES_PATH = "src/test/resources/org/folio/processing/mapping/instance/subjectTypes.json";
+  private static final String DEFAULT_INSTANCE_DATE_TYPES_PATH = "src/test/resources/org/folio/processing/mapping/instance/instanceDateTypes.json";
+
 
   private static final String STUB_FIELD_TYPE_ID = "fe19bae4-da28-472b-be90-d442e2428ead";
   private static final String TXT_INSTANCE_TYPE_ID = "6312d172-f0cf-40f6-b27d-9fa8feaf332f";
@@ -542,6 +547,39 @@ public class InstanceMappingTest {
       assertEquals(expected.getSourceId(), actual.getSourceId());
       assertEquals(expected.getTypeId(), actual.getTypeId());
     });
+  }
+
+  @Test
+  public void testMarcToInstanceWith008Date() throws IOException {
+    MarcReader reader = new MarcStreamReader(new ByteArrayInputStream(TestUtil.readFileFromPath(BIB_WITH_008_DATE).getBytes(StandardCharsets.UTF_8)));
+    JsonObject mappingRules = new JsonObject(TestUtil.readFileFromPath(DEFAULT_MAPPING_RULES_PATH));
+    String rawInstanceDateTypes = TestUtil.readFileFromPath(DEFAULT_INSTANCE_DATE_TYPES_PATH);
+    List<InstanceDateType> instanceDateTypes = List.of(new ObjectMapper().readValue(rawInstanceDateTypes, InstanceDateType[].class));
+
+
+    ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+    List<Instance> mappedInstances = new ArrayList<>();
+    while (reader.hasNext()) {
+      ByteArrayOutputStream os = new ByteArrayOutputStream();
+      MarcJsonWriter writer = new MarcJsonWriter(os);
+      Record targetRecord = reader.next();
+      writer.write(targetRecord);
+      JsonObject marc = new JsonObject(os.toString());
+      Instance instance = mapper.mapRecord(marc, new MappingParameters().withInstanceDateTypes(instanceDateTypes), mappingRules);
+      mappedInstances.add(instance);
+      Validator validator = factory.getValidator();
+      Set<ConstraintViolation<Instance>> violations = validator.validate(instance);
+      assertTrue(violations.isEmpty());
+    }
+    assertFalse(mappedInstances.isEmpty());
+    assertEquals(1, mappedInstances.size());
+
+    Instance mappedInstance = mappedInstances.get(0);
+    assertNotNull(mappedInstance.getId());
+
+    assertEquals("1991", mappedInstances.get(0).getDates().getDate1());
+    assertEquals("0101", mappedInstances.get(0).getDates().getDate2());
+    assertEquals("24a506e8-2a92-4ecc-bd09-ff849321fd5a", mappedInstances.get(0).getDates().getDateTypeId());
   }
 
   @Test